+++ sgtl/graph.py
@@ -98,7 +98,7 @@ class Graph(object):
         :param vertex_set: an iterable collection of vertex indices
         :return: The volume of vertex_set
         """
-        self.check_vert_num(vertex_set)
+        self._check_vert_num(vertex_set)
         return sum([self.degrees[v] for v in vertex_set])
 
     def weight(self, vertex_set_l, vertex_set_r, check_for_overlap=True, sets_are_equal=False):
@@ -117,7 +117,7 @@ class Graph(object):
         :param sets_are_equal: set to ``True`` if the given sets are guaranteed to be equal
         :return: The weight w(L, R)
         """
-        self.check_vert_num(vertex_set_l, vertex_set_r)
+        self._check_vert_num(vertex_set_l, vertex_set_r)
         raw_weight = self.lil_adj_mat[vertex_set_l][:, vertex_set_r].sum()
 
         # If the two sets L and R overlap, we will have double counted any edges inside this overlap.
@@ -143,7 +143,7 @@ class Graph(object):
         :param vertex_set_s: a collection of vertex indices corresponding to the set S
         :return: The conductance :math:`\\phi(S)`
         """
-        self.check_vert_num(vertex_set_s)
+        self._check_vert_num(vertex_set_s)
         return 1 - (2 * self.weight(vertex_set_s, vertex_set_s, sets_are_equal=True)) / self.volume(vertex_set_s)
 
     def bipartiteness(self, vertex_set_l, vertex_set_r):
@@ -159,17 +159,18 @@ class Graph(object):
         :param vertex_set_r: a collection of vertex indices corresponding to the set R
         :return: The bipartiteness ratio \beta(L, R)
         """
-        self.check_vert_num(vertex_set_l, vertex_set_r)
+        self._check_vert_num(vertex_set_l, vertex_set_r)
         return 1 - 2 * self.weight(vertex_set_l, vertex_set_r) / self.volume(vertex_set_l + vertex_set_r)
 
-    def check_vert_num(self, *args):
+    def _check_vert_num(self, *args):
         """
         Check that the number of vertices in a set is not greater than the number of vertices in the graph
         :param *args: the sets to be tested
         """
         for arg in args:
-            if len(arg) > self.num_vertices:
-                raise IndexError("Input vertex set includes indices larger than the number of vertices")
+            for vert in arg:
+                if vert >= self.num_vertices:
+                    raise IndexError("Input vertex set includes indices larger than the number of vertices.")
 
 
 def complete_graph(n: int) -> Graph: